# shared/openapi/api-gateway.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CogniBoiler API Gateway",
    "description": "REST and WebSocket gateway of the CogniBoiler digital twin: sign-in and sessions, plant state, PLC commands, simulation control, alarms, history, audit and user administration. Errors are Problem Details (RFC 9457).",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/history": {
      "get": {
        "tags": [
          "history"
        ],
        "summary": "Get History",
        "description": "Telemetry over a bounded range at a resolution chosen from its length.\n\nThe range is split into standard windows (1 s … 1 day) so that it fits in `limit`\npoints; each point is the mean of its window.",
        "operationId": "get_history_api_v1_history_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "measurement",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(boiler_sensors|turbine_sensors|plant_status)$",
              "default": "boiler_sensors",
              "title": "Measurement"
            }
          },
          {
            "name": "start_ms",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "[UTC epoch ms]",
              "title": "Start Ms"
            },
            "description": "[UTC epoch ms]"
          },
          {
            "name": "end_ms",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "[UTC epoch ms]",
              "title": "End Ms"
            },
            "description": "[UTC epoch ms]"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 2000,
              "minimum": 1,
              "description": "Maximum number of points.",
              "default": 200,
              "title": "Limit"
            },
            "description": "Maximum number of points."
          },
          {
            "name": "fields",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated field names; all when omitted.",
              "title": "Fields"
            },
            "description": "Comma-separated field names; all when omitted."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/alarms": {
      "get": {
        "tags": [
          "alarms"
        ],
        "summary": "List Alarms",
        "description": "Recent alarms; with active_only, open alarms with critical and unacknowledged first.",
        "operationId": "list_alarms_api_v1_alarms_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "active_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Only open alarms: not both cleared and acknowledged.",
              "default": false,
              "title": "Active Only"
            },
            "description": "Only open alarms: not both cleared and acknowledged."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AlarmResponse"
                  },
                  "title": "Response List Alarms Api V1 Alarms Get"
                }
              }
            }
          },
          "503": {
            "description": "upstream.unavailable: the upstream service (named in `service`) failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "504": {
            "description": "upstream.timeout: the upstream service did not answer in time; a command may still have been carried out.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/alarms/history": {
      "get": {
        "tags": [
          "alarms"
        ],
        "summary": "Alarm History",
        "description": "Alarm history with filters, newest first, paged.",
        "operationId": "alarm_history_api_v1_alarms_history_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "severity",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "warning",
                    "critical"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severity"
            }
          },
          {
            "name": "parameter",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 128
                },
                {
                  "type": "null"
                }
              ],
              "title": "Parameter"
            }
          },
          {
            "name": "from_ms",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Raised at or after [UTC ms].",
              "default": 0,
              "title": "From Ms"
            },
            "description": "Raised at or after [UTC ms]."
          },
          {
            "name": "to_ms",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Raised at or before [UTC ms].",
              "default": 0,
              "title": "To Ms"
            },
            "description": "Raised at or before [UTC ms]."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlarmPageResponse"
                }
              }
            }
          },
          "503": {
            "description": "upstream.unavailable: the upstream service (named in `service`) failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "504": {
            "description": "upstream.timeout: the upstream service did not answer in time; a command may still have been carried out.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/alarms/ack-all": {
      "post": {
        "tags": [
          "alarms"
        ],
        "summary": "Acknowledge All",
        "description": "Acknowledge every unacknowledged alarm, optionally of one severity.",
        "operationId": "acknowledge_all_api_v1_alarms_ack_all_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcknowledgeAllRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcknowledgeResponse"
                }
              }
            }
          },
          "503": {
            "description": "upstream.unavailable: the upstream service (named in `service`) failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "504": {
            "description": "upstream.timeout: the upstream service did not answer in time; a command may still have been carried out.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/alarms/{alarm_id}": {
      "get": {
        "tags": [
          "alarms"
        ],
        "summary": "Get Alarm",
        "description": "One alarm with every state change it went through.",
        "operationId": "get_alarm_api_v1_alarms__alarm_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "alarm_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Alarm Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlarmDetailResponse"
                }
              }
            }
          },
          "503": {
            "description": "upstream.unavailable: the upstream service (named in `service`) failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "504": {
            "description": "upstream.timeout: the upstream service did not answer in time; a command may still have been carried out.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/alarms/{alarm_id}/ack": {
      "post": {
        "tags": [
          "alarms"
        ],
        "summary": "Acknowledge Alarm",
        "description": "Acknowledge one alarm; refused if it is already acknowledged.",
        "operationId": "acknowledge_alarm_api_v1_alarms__alarm_id__ack_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "alarm_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Alarm Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcknowledgeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcknowledgeResponse"
                }
              }
            }
          },
          "503": {
            "description": "upstream.unavailable: the upstream service (named in `service`) failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "504": {
            "description": "upstream.timeout: the upstream service did not answer in time; a command may still have been carried out.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AcknowledgeAllRequest": {
        "properties": {
          "comment": {
            "type": "string",
            "maxLength": 500,
            "title": "Comment",
            "default": ""
          },
          "severity": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "warning",
                  "critical"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Severity",
            "description": "Only alarms of this severity; all when omitted."
          }
        },
        "type": "object",
        "title": "AcknowledgeAllRequest",
        "description": "Request body for POST /api/v1/alarms/ack-all."
      },
      "AcknowledgeRequest": {
        "properties": {
          "comment": {
            "type": "string",
            "maxLength": 500,
            "title": "Comment",
            "default": ""
          }
        },
        "type": "object",
        "title": "AcknowledgeRequest",
        "description": "Request body for POST /api/v1/alarms/{alarm_id}/ack."
      },
      "AcknowledgeResponse": {
        "properties": {
          "accepted": {
            "type": "boolean",
            "title": "Accepted"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "timestamp_ms": {
            "type": "integer",
            "title": "Timestamp Ms"
          },
          "alarms": {
            "items": {
              "$ref": "#/components/schemas/AlarmResponse"
            },
            "type": "array",
            "title": "Alarms"
          }
        },
        "type": "object",
        "required": [
          "accepted",
          "reason",
          "timestamp_ms",
          "alarms"
        ],
        "title": "AcknowledgeResponse",
        "description": "Result of an acknowledgement."
      },
...
      "AlarmDetailResponse": {
        "properties": {
          "alarm": {
            "$ref": "#/components/schemas/AlarmResponse"
          },
          "transitions": {
            "items": {
              "$ref": "#/components/schemas/AlarmTransitionResponse"
            },
            "type": "array",
            "title": "Transitions"
          }
        },
        "type": "object",
        "required": [
          "alarm",
          "transitions"
        ],
        "title": "AlarmDetailResponse",
        "description": "An alarm with its transitions, oldest first."
      },
      "AlarmPageResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AlarmResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "AlarmPageResponse",
        "description": "A page of alarm history."
      },
      "AlarmResponse": {
        "properties": {
          "alarm_id": {
            "type": "string",
            "title": "Alarm Id",
            "description": "The alarm id as a string."
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "key": {
            "type": "string",
            "title": "Key",
            "description": "Condition identity; one open alarm per key."
          },
          "source_service": {
            "type": "string",
            "title": "Source Service"
          },
          "severity": {
            "type": "string",
            "title": "Severity"
          },
          "parameter": {
            "type": "string",
            "title": "Parameter"
          },
          "direction": {
            "type": "string",
            "title": "Direction"
          },
          "unit": {
            "type": "string",
            "title": "Unit"
          },
          "state": {
            "type": "string",
            "enum": [
              "ACTIVE_UNACK",
              "ACTIVE_ACK",
              "CLEARED_UNACK",
              "CLEARED"
            ],
            "title": "State"
          },
          "value": {
            "type": "number",
            "title": "Value",
            "description": "Latest value, in `unit`."
          },
          "threshold": {
            "type": "number",
            "title": "Threshold"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "topic": {
            "type": "string",
            "title": "Topic"
          },
          "occurred_at_ms": {
            "type": "integer",
            "title": "Occurred At Ms",
            "description": "Same as raised_at_ms."
          },
          "raised_at_ms": {
            "type": "integer",
            "title": "Raised At Ms"
          },
          "cleared_at_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cleared At Ms"
          },
          "acknowledged": {
            "type": "boolean",
            "title": "Acknowledged"
          },
          "acknowledged_at_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acknowledged At Ms"
          },
          "acknowledged_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acknowledged By"
          },
          "ack_comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ack Comment"
          },
          "cleared": {
            "type": "boolean",
            "title": "Cleared",
            "description": "The condition has ended."
          },
          "occurrence_count": {
            "type": "integer",
            "title": "Occurrence Count"
          },
          "updated_at_ms": {
            "type": "integer",
            "title": "Updated At Ms"
          }
        },
        "type": "object",
        "required": [
          "alarm_id",
          "id",
          "key",
          "source_service",
          "severity",
          "parameter",
          "direction",
          "unit",
          "state",
          "value",
          "threshold",
          "action",
          "message",
          "topic",
          "occurred_at_ms",
          "raised_at_ms",
          "cleared_at_ms",
          "acknowledged",
          "acknowledged_at_ms",
          "acknowledged_by",
          "ack_comment",
          "cleared",
          "occurrence_count",
          "updated_at_ms"
        ],
        "title": "AlarmResponse",
        "description": "An alarm with its lifecycle.\n\n`alarm_id`, `occurred_at_ms`, `acknowledged` and `cleared` keep their earlier\nmeaning; `state` is the full lifecycle state."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HistoryResponse": {
        "properties": {
          "measurement": {
            "type": "string",
            "title": "Measurement"
          },
          "start_ms": {
            "type": "integer",
            "title": "Start Ms",
            "description": "Range start [UTC epoch ms]."
          },
          "end_ms": {
            "type": "integer",
            "title": "End Ms",
            "description": "Range end [UTC epoch ms]."
          },
          "window_s": {
            "type": "integer",
            "title": "Window S",
            "description": "Each point is the mean over a window this long [s]."
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/HistoryPointResponse"
            },
            "type": "array",
            "title": "Points"
          }
        },
        "type": "object",
        "required": [
          "measurement",
          "start_ms",
          "end_ms",
          "window_s",
          "points"
        ],
        "title": "HistoryResponse",
        "description": "Historical telemetry over a bounded range at an automatic resolution."
      },
...
      "ProblemDetails": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "integer",
            "title": "Status"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "instance": {
            "type": "string",
            "title": "Instance"
          },
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Stable, machine-readable error code."
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "code"
        ],
        "title": "ProblemDetails",
        "description": "An error body (RFC 9457); members beyond these depend on the code."
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}